requirements.txt
django==4.1.6
python-decouple==3.7
djangorestframework==3.14.0
django-model-utils==4.3.1
django-cors-headers==3.13.0
drf-spectacular==0.25.1
aiogram==2.23.1
requests==2.31.0
